mqtt_system/requirements.txt
# =====================================================
# 波浪监测系统 - Python依赖
# =====================================================

# MQTT客户端
paho-mqtt>=1.6.1

# 配置文件解析
PyYAML>=6.0

# 数据库
psycopg2-binary>=2.9.5

# 科学计算
numpy>=1.24.0
scipy>=1.10.0
pandas>=1.5.0

# Web框架
Flask>=2.3.0
flask-cors>=4.0.0
flask-sock>=0.7.0

# 串口通信
pyserial>=3.5

# 时间处理
python-dateutil>=2.8.2

# 可选：数据分析和可视化
# matplotlib>=3.7.0
# seaborn>=0.12.0

# 可选：netCDF支持（如需要xarray）
# xarray>=2023.1.0
# netCDF4>=1.6.0

# 可选：MHKiT（专业波浪分析）
# mhkit>=0.6.0

# 开发和测试
# pytest>=7.2.0
# pytest-cov>=4.0.0
